feat: move type to local private npm package
