+++ lib/podcast-types/index.d.ts
@@ -1,0 +1,5 @@
+export * from './PodcastDetails';
+export * from './PodcastEntry';
+export * from './PodcastEpisode';
+export * from './PodcastFeed';
+export * from './PodcastImage';
