+++ lib/podcast-types/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "podcast-types",
+  "version": "1.0.0",
+  "main": "index.d.ts",
+  "types": "index.d.ts",
+  "files": [
+    "*.d.ts"
+  ],
+  "private": true
+}
